Fixing recent Windows 10 issue
After a recent Windows 10 update Solace stopped instantly changing the
colors of the command prompt, instead forcing me to run "COLOR 01",
"COLOR 07" to change the color to something else then back, and *also*
run "CLS" to clear the screen before the new color palette could be seen
everywhere. This change causes the old behavior to return in Solace by
essentially doing the color change automatically and changing the color
settings twice, which causes them to take effect, which is of course a
kludge but it works.